requirements.txt
pytz~=2021.1
PyYAML~=5.4.1
plotly~=4.14.3